+++ test/formatters/samplecode/lisp-indent/in.el
@@ -1,0 +1,5 @@
+;; -*- indent-tabs-mode: nil -*-
+(if (and (< 3 5)
+                (= 1 1))
+                                 (message "true")
+(message "false"))